+++ src/test/java/step3/domain/movement/MovementConditionProviderTest.java
@@ -1,4 +1,4 @@
-package step3.service;
+package step3.domain.movement;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
